Fixing size handling
When images have different orientations (portrait / landscape) usualy the resizing to eg. 1200x1200 the long side should fit into this values and the short side should be resized according to fit aspectRatio.

Actually the task ignores images, when one side of the original image is smaller, than the according sizeTo-value.
e.g. the original image is 1500x500 and it should fit into 1200x1200 actually the task generates an image with 1500x500 instead of 1200x400.

This Bug is fixed with this changes.